Separate pin tests into smaller files
